made correction

+++ src/com/company/Main.java
@@ -6,7 +6,7 @@ public class Main {
     public static void main(String[] args) {
         System.out.println("Enter Radius of Circle: ");
         Scanner scanner = new Scanner(System.in);
-        int radius = scanner.nextInt();
+        double radius = scanner.nextInt();
 
         Circle circle = new Circle();
         circle.circumference(radius);
